move _convert_to_int to demean.pyÄ '
